feat: extract ta-plan crate + PlanStore trait (v0.17.11.1)
Extracts PLAN.md schema/parse/query/history logic out of
apps/ta-cli/src/commands/plan.rs into a new crates/ta-plan library
crate, matching ta-goal's existing position below ta-daemon/
ta-mcp-gateway/ta-cli. Pure extraction, no behavior change: plan.rs
re-exports via `pub use ta_plan::{...}` instead of duplicating logic.

Adds the PlanStore trait (phase layer + goal layer) with FilePlanStore
as its sole implementation, wrapping the extracted ta-plan functions
and GoalRunStore. Deliberately synchronous (documented rationale in
the module doc comment) since every current caller and implementation
is a synchronous filesystem operation.

ta-daemon's independently-implemented duplicate PLAN.md parser is kept
deliberately (different response shape, live-used by Studio) with the
divergence now documented in-code instead of left silent.

Verified via a real-PLAN.md regression test plus a full
cargo test -p ta-cli --bin ta pass (1625 tests, 0 failed).

Remaining gaps (call-site wiring, Done/Deferred transitions, a second
PlanStore impl) tracked as v0.17.11.1.2 in PLAN.md.
