+++ apps/ta-cli/src/commands/plan.rs
@@ -21,14 +21,12 @@
 // `ta pr apply` auto-updates PLAN.md when a goal with --phase completes.
 
 use std::cmp::Reverse;
-use std::fmt;
 use std::path::Path;
 
 use clap::Subcommand;
 use regex::Regex;
-use ta_goal::{extract_human_review_items, HumanReviewStore};
+use ta_goal::HumanReviewStore;
 use ta_mcp_gateway::GatewayConfig;
-use ta_submit::WorkflowConfig as TaWorkflowConfig;
 
 #[derive(Debug, Subcommand)]
 pub enum PlanCommands {
@@ -689,548 +687,16 @@ pub fn execute(cmd: &PlanCommands, config: &GatewayConfig) -> anyhow::Result<()>
     }
 }
 
-// ── Data model ───────────────────────────────────────────────────
-
-/// Status of a plan phase.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum PlanStatus {
-    Pending,
-    InProgress,
-    Done,
-    /// Deferred phases are excluded from "next pending" but still appear in the checklist.
-    Deferred,
-}
-
-impl PlanStatus {
-    /// Returns true if this phase can be dispatched as new work.
-    ///
-    /// `InProgress` is NOT actionable — it means the phase is already claimed by a running
-    /// goal and must be skipped by `find_next_pending`. Only `Pending` phases are eligible
-    /// for new dispatch. (v0.15.24.2: fixed from `Pending | InProgress` to `Pending` only.)
-    pub fn is_actionable(&self) -> bool {
-        matches!(self, PlanStatus::Pending)
-    }
-
-    /// Returns true if the transition from `self` to `to` is a legal state-machine move.
-    ///
-    /// Legal transitions:
-    ///   `pending    → in_progress`  (claim: ta run)
-    ///   `in_progress → done`         (complete: ta draft apply)
-    ///   `in_progress → pending`      (reset: ta draft deny or ta goal delete)
-    ///
-    /// Everything else is illegal.
-    pub fn is_valid_transition_to(&self, to: &PlanStatus) -> bool {
-        matches!(
-            (self, to),
-            (PlanStatus::Pending, PlanStatus::InProgress)
-                | (PlanStatus::InProgress, PlanStatus::Done)
-                | (PlanStatus::InProgress, PlanStatus::Pending)
-        )
-    }
-}
-
-impl fmt::Display for PlanStatus {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            PlanStatus::Pending => write!(f, "pending"),
-            PlanStatus::InProgress => write!(f, "in_progress"),
-            PlanStatus::Done => write!(f, "done"),
-            PlanStatus::Deferred => write!(f, "deferred"),
-        }
-    }
-}
-
-/// A parsed plan phase from PLAN.md.
-#[derive(Debug, Clone)]
-pub struct PlanPhase {
-    /// Phase identifier (e.g., "0", "4b", "4a.1").
-    pub id: String,
-    /// Human-readable title (e.g., "Per-Artifact Review Model").
-    pub title: String,
-    /// Current status.
-    pub status: PlanStatus,
-    /// Explicit dependencies declared via `<!-- depends_on: v0.13.17.3 -->` comment
-    /// (v0.14.3) or, far more commonly in practice, a `**Depends on**: v0.13.17.3
-    /// (explanation), v0.14.1 (...)` prose line (parsed since v0.17.0.12.34 —
-    /// see `find_depends_on_in_lookahead`). Parenthetical explanations are
-    /// stripped; only the leading phase-id-shaped token from each entry is kept.
-    pub depends_on: Vec<String>,
-    /// Items from the `#### Human Review` subsection of this phase (v0.15.14.1).
-    ///
-    /// These items require a human to verify or sign off — agents must not check them.
-    pub human_review_items: Vec<String>,
-    /// Declared API surfaces this phase's work touches (v0.17.0.12.34), from a
-    /// `**API impact**: adds Foo::bar; modifies Baz::qux` prose line. Used by
-    /// the dependency-wave planner (`candidate_waves`) to downgrade two
-    /// otherwise-independent phases to sequential waves when they declare
-    /// touching the same API surface — a conflict class plain file-overlap
-    /// can't catch.
-    pub api_impact: Vec<String>,
-}
-
-// ── Schema-driven parsing ────────────────────────────────────────
-
-/// A single phase-header pattern in the schema.
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct PhasePattern {
-    /// Regex with capturing groups: group 1 = phase ID, group 2 (optional) = title.
-    pub regex: String,
-    /// Human-readable label for what this pattern captures (informational only).
-    #[serde(default)]
-    pub id_capture: String,
-}
-
-/// Schema describing how to parse a project's plan document.
-/// Loaded from `.ta/plan-schema.yaml`. If absent, the built-in default is used.
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct PlanSchema {
-    /// Path to the plan file, relative to project root (default: "PLAN.md").
-    #[serde(default = "default_source")]
-    pub source: String,
-    /// One or more header patterns for phase detection (evaluated in order, first match wins).
-    pub phase_patterns: Vec<PhasePattern>,
-    /// Regex with one capture group that extracts the status value.
-    pub status_marker: String,
-    /// Recognized status values. Anything not in this list maps to Pending.
-    #[serde(default = "default_statuses")]
-    pub statuses: Vec<String>,
-    /// Directories to search when resolving document paths in `ta plan from`.
-    /// Relative to the project root. Searched in order; first match wins.
-    /// If omitted, uses sensible defaults (docs/, spec/, design/, etc.).
-    #[serde(default = "default_doc_search_dirs")]
-    pub doc_search_dirs: Vec<String>,
-}
-
-fn default_source() -> String {
-    "PLAN.md".to_string()
-}
-
-fn default_statuses() -> Vec<String> {
-    vec![
-        "done".to_string(),
-        "in_progress".to_string(),
-        "pending".to_string(),
-        "deferred".to_string(),
-    ]
-}
-
-fn default_doc_search_dirs() -> Vec<String> {
-    vec![
-        ".".to_string(),
-        "docs".to_string(),
-        "doc".to_string(),
-        "documentation".to_string(),
-        "specs".to_string(),
-        "spec".to_string(),
-        "design".to_string(),
-        "rfcs".to_string(),
-        "rfc".to_string(),
-        "planning".to_string(),
-        "plans".to_string(),
-        "requirements".to_string(),
-        ".ta".to_string(),
-    ]
-}
-
-impl PlanSchema {
-    /// The built-in default schema — matches the current PLAN.md format.
-    /// Used when no `.ta/plan-schema.yaml` is present.
-    pub fn default_schema() -> Self {
-        PlanSchema {
-            source: "PLAN.md".to_string(),
-            phase_patterns: vec![
-                PhasePattern {
-                    // Matches: "## Phase 4b — Title" and "## Phase 4a.1 — Title"
-                    regex: r"^##\s+Phase[\s\u{a0}]+([0-9a-z.]+)\s+[—\-]\s+(.+)$".to_string(),
-                    id_capture: "phase_number".to_string(),
-                },
-                PhasePattern {
-                    // Matches: "### v0.3.1 — Title" or "### v0.3.1.1 — Title"
-                    regex: r"^###\s+(v[\d.]+[a-z]?)\s+[—\-]\s+(.+)$".to_string(),
-                    id_capture: "version_number".to_string(),
-                },
-            ],
-            status_marker: r"<!--\s*status:\s*(\w+)\s*-->".to_string(),
-            statuses: default_statuses(),
-            doc_search_dirs: default_doc_search_dirs(),
-        }
-    }
-
-    /// Load schema from `.ta/plan-schema.yaml`, falling back to `default_schema()`.
-    pub fn load_or_default(project_root: &Path) -> Self {
-        let schema_path = project_root.join(".ta/plan-schema.yaml");
-        if schema_path.exists() {
-            if let Ok(content) = std::fs::read_to_string(&schema_path) {
-                if let Ok(schema) = serde_yaml::from_str::<PlanSchema>(&content) {
-                    return schema;
-                }
-                eprintln!("Warning: failed to parse .ta/plan-schema.yaml — using default schema");
-            }
-        }
-        Self::default_schema()
-    }
-
-    /// Serialize to YAML string.
-    pub fn to_yaml(&self) -> anyhow::Result<String> {
-        Ok(serde_yaml::to_string(self)?)
-    }
-}
-
-// ── Parsing ──────────────────────────────────────────────────────
-
-/// Parse plan content using a provided schema.
-///
-/// Each `phase_patterns` regex is tested against each line.
-/// The first match wins. The regex must have:
-///   - Group 1: phase ID (e.g., "4b", "v0.3.1")
-///   - Group 2 (optional): phase title
-///
-/// The status marker regex is tested against the next non-empty line.
-pub fn parse_plan_with_schema(content: &str, schema: &PlanSchema) -> Vec<PlanPhase> {
-    // Pre-compile all regexes. Silently skip invalid ones.
-    let compiled_patterns: Vec<Regex> = schema
-        .phase_patterns
-        .iter()
-        .filter_map(|p| Regex::new(&p.regex).ok())
-        .collect();
-
-    let status_re = match Regex::new(&schema.status_marker) {
-        Ok(r) => r,
-        Err(_) => return vec![],
-    };
-
-    let lines: Vec<&str> = content.lines().collect();
-    let mut phases = Vec::new();
-    let mut i = 0;
-
-    while i < lines.len() {
-        let line = lines[i].trim();
-
-        for pattern in &compiled_patterns {
-            if let Some(caps) = pattern.captures(line) {
-                let id = caps
-                    .get(1)
-                    .map(|m| m.as_str().trim().to_string())
-                    .unwrap_or_default();
-                let title = caps
-                    .get(2)
-                    .map(|m| m.as_str().trim().to_string())
-                    .unwrap_or_default();
-
-                if id.is_empty() {
-                    break;
-                }
-
-                // Strip trailing markup from title (e.g. "*(release)*").
-                let title = title.trim_end_matches(['*', '(', ')']).trim().to_string();
-
-                let status = find_status_in_lookahead(&lines, i + 1, &status_re);
-                let depends_on = find_depends_on_in_lookahead(&lines, i + 1);
-                let api_impact = find_api_impact_in_lookahead(&lines, i + 1);
-                let human_review_items = extract_human_review_items(content, &id, &title);
-                phases.push(PlanPhase {
-                    id,
-                    title,
-                    status,
-                    depends_on,
-                    human_review_items,
-                    api_impact,
-                });
-                break; // First pattern match wins.
-            }
-        }
-
-        i += 1;
-    }
-
-    phases
-}
-
-/// Compare phase IDs, normalizing the optional `v` prefix.
-/// e.g., "v0.4.0" matches "0.4.0", "4b" matches "4b".
-pub fn phase_ids_match(parsed_id: &str, phase_id: &str) -> bool {
-    if parsed_id == phase_id {
-        return true;
-    }
-    let norm_parsed = parsed_id.strip_prefix('v').unwrap_or(parsed_id);
-    let norm_phase = phase_id.strip_prefix('v').unwrap_or(phase_id);
-    norm_parsed == norm_phase
-}
-
-/// Look ahead from `start` for a status marker comment.
-/// Skips blank lines (up to 3) so that a blank line between a phase heading
-/// and its `<!-- status: ... -->` marker does not cause it to read as Pending.
-/// Stops immediately on the first non-blank, non-status line.
-fn find_status_in_lookahead(lines: &[&str], start: usize, status_re: &Regex) -> PlanStatus {
-    let mut skipped = 0;
-    let mut i = start;
-    while i < lines.len() && skipped <= 3 {
-        let line = lines[i].trim();
-        if line.is_empty() {
-            skipped += 1;
-            i += 1;
-            continue;
-        }
-        if let Some(caps) = status_re.captures(line) {
-            let status_str = caps.get(1).map(|m| m.as_str().trim()).unwrap_or("");
-            return parse_status_str(status_str);
-        }
-        // First non-blank line that isn't a status marker — stop scanning.
-        break;
-    }
-    PlanStatus::Pending
-}
-
-/// Look ahead from `start` for a dependency declaration, in either of two
-/// forms actually seen in PLAN.md:
-///   - `<!-- depends_on: v0.13.17.3, v0.14.1 -->` (v0.14.3, a plain comma
-///     list, rarely used in practice — 1 occurrence across the whole doc).
-///   - `**Depends on**: v0.13.17.3 (explanation), v0.14.1 (...)` — a bold
-///     prose line (v0.17.0.12.34), the format actually used ~125 times.
-///     Parenthetical explanations are stripped; only the leading
-///     phase-id-shaped token from each comma-separated entry is kept, and
-///     entries that don't start with one (e.g. "Meridian `suggest` command
-///     (v0.1.x)") are silently skipped rather than guessed at.
-///
-/// Scans up to 5 lines ahead, stopping if another phase header is detected.
-fn find_depends_on_in_lookahead(lines: &[&str], start: usize) -> Vec<String> {
-    let dep_comment_re = match Regex::new(r"<!--\s*depends_on:\s*([^>]+?)\s*-->") {
-        Ok(r) => r,
-        Err(_) => return vec![],
-    };
-    let dep_prose_re = match Regex::new(r"^\*\*Depends [Oo]n\*\*:\s*(.+)$") {
-        Ok(r) => r,
-        Err(_) => return vec![],
-    };
-    // Phase header patterns to detect the next phase boundary.
-    let header_re = match Regex::new(r"^(?:##\s+Phase|###\s+v[\d.]+[a-z]?\s+[—\-])") {
-        Ok(r) => r,
-        Err(_) => return vec![],
-    };
-    let limit = std::cmp::min(start + 5, lines.len());
-    for (offset, line) in lines[start..limit].iter().enumerate() {
-        let line = line.trim();
-        // Stop if we've hit the next phase header (but not on the first lookahead line).
-        if offset > 0 && header_re.is_match(line) {
-            break;
-        }
-        if let Some(caps) = dep_comment_re.captures(line) {
-            let raw = caps.get(1).map(|m| m.as_str()).unwrap_or("");
-            return raw
-                .split(',')
-                .map(|s| s.trim().to_string())
-                .filter(|s| !s.is_empty())
-                .collect();
-        }
-        if let Some(caps) = dep_prose_re.captures(line) {
-            let raw = caps.get(1).map(|m| m.as_str()).unwrap_or("");
-            return extract_leading_id_tokens(raw);
-        }
-    }
-    vec![]
-}
-
-/// Look ahead from `start` for a `**API impact**: adds Foo::bar; modifies
-/// Baz::qux` prose line (v0.17.0.12.34). Entries are semicolon-separated
-/// free-text tokens (not phase IDs) — trimmed verbatim, no id extraction.
-/// Scans up to 5 lines ahead, stopping if another phase header is detected.
-fn find_api_impact_in_lookahead(lines: &[&str], start: usize) -> Vec<String> {
-    let impact_re = match Regex::new(r"^\*\*API [Ii]mpact\*\*:\s*(.+)$") {
-        Ok(r) => r,
-        Err(_) => return vec![],
-    };
-    let header_re = match Regex::new(r"^(?:##\s+Phase|###\s+v[\d.]+[a-z]?\s+[—\-])") {
-        Ok(r) => r,
-        Err(_) => return vec![],
-    };
-    let limit = std::cmp::min(start + 5, lines.len());
-    for (offset, line) in lines[start..limit].iter().enumerate() {
-        let line = line.trim();
-        if offset > 0 && header_re.is_match(line) {
-            break;
-        }
-        if let Some(caps) = impact_re.captures(line) {
-            let raw = caps.get(1).map(|m| m.as_str()).unwrap_or("");
-            return split_top_level(raw, ';')
-                .into_iter()
-                .map(|s| s.trim().to_string())
-                .filter(|s| !s.is_empty())
-                .collect();
-        }
-    }
-    vec![]
-}
-
-/// Split `s` on top-level occurrences of `sep`, treating `(...)` spans as
-/// opaque so a comma or semicolon inside a parenthetical explanation doesn't
-/// split an entry in half.
-fn split_top_level(s: &str, sep: char) -> Vec<String> {
-    let mut parts = Vec::new();
-    let mut depth = 0i32;
-    let mut current = String::new();
-    for ch in s.chars() {
-        match ch {
-            '(' => {
-                depth += 1;
-                current.push(ch);
-            }
-            ')' => {
-                depth -= 1;
-                current.push(ch);
-            }
-            c if c == sep && depth <= 0 => {
-                parts.push(current.trim().to_string());
-                current.clear();
-            }
-            c => current.push(c),
-        }
-    }
-    if !current.trim().is_empty() {
-        parts.push(current.trim().to_string());
-    }
-    parts
-}
-
-/// From a `**Depends on**:` prose line's captured text, extract the leading
-/// phase-id-shaped token of each top-level comma-separated entry, dropping
-/// any parenthetical explanation. Entries that don't start with an
-/// id-shaped token (e.g. "Meridian `suggest` command (v0.1.x)", or "None")
-/// are silently skipped — this is a conservative extraction, not a guess.
-fn extract_leading_id_tokens(raw: &str) -> Vec<String> {
-    let id_re = match Regex::new(r"^(v?\d[\da-z.]*)") {
-        Ok(r) => r,
-        Err(_) => return vec![],
-    };
-    split_top_level(raw, ',')
-        .iter()
-        .filter_map(|entry| {
-            id_re
-                .captures(entry.trim())
-                .and_then(|c| c.get(1))
-                .map(|m| m.as_str().trim_end_matches('.').to_string())
-        })
-        .collect()
-}
-
-fn parse_status_str(s: &str) -> PlanStatus {
-    match s {
-        "done" => PlanStatus::Done,
-        "in_progress" => PlanStatus::InProgress,
-        "deferred" => PlanStatus::Deferred,
-        _ => PlanStatus::Pending,
-    }
-}
-
-/// Parse PLAN.md content into a list of phases (using the default schema).
-///
-/// This is the backward-compatible entry point used by existing code.
-pub fn parse_plan(content: &str) -> Vec<PlanPhase> {
-    parse_plan_with_schema(content, &PlanSchema::default_schema())
-}
-
-/// Update a phase's status in PLAN.md content. Returns the new content.
-///
-/// Finds the phase by ID using the default schema's patterns
-/// and replaces its status marker.
-pub fn update_phase_status(content: &str, phase_id: &str, new_status: PlanStatus) -> String {
-    update_phase_status_with_schema(content, phase_id, new_status, &PlanSchema::default_schema())
-}
-
-/// Update a phase's status using a provided schema.
-pub fn update_phase_status_with_schema(
-    content: &str,
-    phase_id: &str,
-    new_status: PlanStatus,
-    schema: &PlanSchema,
-) -> String {
-    let compiled_patterns: Vec<Regex> = schema
-        .phase_patterns
-        .iter()
-        .filter_map(|p| Regex::new(&p.regex).ok())
-        .collect();
-
-    let status_re = match Regex::new(&schema.status_marker) {
-        Ok(r) => r,
-        Err(_) => return content.to_string(),
-    };
-
-    let lines: Vec<&str> = content.lines().collect();
-    let mut result = Vec::with_capacity(lines.len());
-    let mut i = 0;
-
-    while i < lines.len() {
-        let line = lines[i];
-        let trimmed = line.trim();
-
-        // Check if this line is the target phase header.
-        // Normalize comparison: "v0.4.0" matches "0.4.0" and vice versa.
-        let mut is_target = false;
-        for pattern in &compiled_patterns {
-            if let Some(caps) = pattern.captures(trimmed) {
-                if let Some(id_match) = caps.get(1) {
-                    let parsed_id = id_match.as_str().trim();
-                    if phase_ids_match(parsed_id, phase_id) {
-                        is_target = true;
-                        break;
-                    }
-                }
-            }
-        }
-
-        result.push(line.to_string());
-
-        // If this is the target phase, find and replace the status marker,
-        // skipping over blank lines (up to 3) between the header and the marker.
-        if is_target {
-            let mut j = i + 1;
-            let mut blank_count = 0;
-            while j < lines.len() && blank_count <= 3 {
-                let next = lines[j].trim();
-                if next.is_empty() {
-                    blank_count += 1;
-                    j += 1;
-                    continue;
-                }
-                if status_re.is_match(next) {
-                    // Emit the blank lines we skipped, then the replacement marker.
-                    for blank_line in &lines[(i + 1)..j] {
-                        result.push(blank_line.to_string());
-                    }
-                    result.push(format!("<!-- status: {} -->", new_status));
-                    i = j + 1;
-                    break;
-                }
-                // Non-blank, non-status line — no marker found; leave as-is.
-                break;
-            }
-            if i == j + 1 {
-                continue;
-            }
-        }
-
-        i += 1;
-    }
-
-    let mut out = result.join("\n");
-    // Preserve trailing newline: `str::lines()` strips it, join() doesn't restore it.
-    if content.ends_with('\n') {
-        out.push('\n');
-    }
-    out
-}
-
-/// Read and parse PLAN.md from a project directory.
-///
-/// Loads `.ta/plan-schema.yaml` if present, otherwise uses the default schema.
-pub fn load_plan(project_root: &Path) -> anyhow::Result<Vec<PlanPhase>> {
-    let schema = PlanSchema::load_or_default(project_root);
-    let plan_path = project_root.join(&schema.source);
-    if !plan_path.exists() {
-        anyhow::bail!("No {} found in {}", schema.source, project_root.display());
-    }
-    let content = std::fs::read_to_string(&plan_path)?;
-    Ok(parse_plan_with_schema(&content, &schema))
-}
+// ── Data model + schema-driven parsing (v0.17.11.1: extracted to ta-plan) ──
+pub use ta_plan::{
+    load_plan, parse_plan, parse_plan_with_schema, phase_ids_match, update_phase_status,
+    update_phase_status_with_schema, PhasePattern, PlanPhase, PlanSchema, PlanStatus,
+};
+// Only used by this module's own tests (constructing a custom PlanSchema
+// without going through PlanSchema::default_schema()) — cfg-gated so a
+// non-test build doesn't warn about an unused import.
+#[cfg(test)]
+use ta_plan::{default_doc_search_dirs, default_statuses};
 
 /// Format a plan phase list as a checklist for CLAUDE.md injection.
 pub fn format_plan_checklist(phases: &[PlanPhase], current_phase: Option<&str>) -> String {
@@ -1392,197 +858,27 @@ pub fn format_plan_checklist_windowed(
 
     lines.join("\n")
 }
+// v0.17.11.1: extracted to ta-plan.
+pub use ta_plan::{find_in_progress, find_next_pending};
+// v0.17.11.1: extracted to ta-plan.
+pub use ta_plan::{mark_phase_in_source, record_history};
 
-/// Find the next actionable phase after the given phase ID.
-///
-/// Skips phases marked as `Deferred`, `Done`, or `InProgress` — only
-/// returns `Pending` phases. `InProgress` means the phase is already
-/// claimed by a running goal; returning it again would cause duplicate
-/// dispatch. If `after_phase` is None, returns the first pending phase.
-pub fn find_next_pending<'a>(
-    phases: &'a [PlanPhase],
-    after_phase: Option<&str>,
-) -> Option<&'a PlanPhase> {
-    if let Some(after) = after_phase {
-        // Find the current phase's position and search forward from there.
-        if let Some(idx) = phases.iter().position(|p| phase_ids_match(&p.id, after)) {
-            // Search forward from the phase after the current one.
-            if let Some(next) = phases[idx + 1..].iter().find(|p| p.status.is_actionable()) {
-                return Some(next);
-            }
-        }
-        // Phase not found or no actionable phases after it — don't fall back to
-        // the beginning (which would suggest unrelated earlier phases like v0.1).
-        None
-    } else {
-        phases.iter().find(|p| p.status.is_actionable())
-    }
-}
-
-/// Find the first `InProgress` phase.
-///
-/// Used for status introspection, resume flows, and claim checks. Not for
-/// dispatch decisions — use `find_next_pending` for those.
-pub fn find_in_progress(phases: &[PlanPhase]) -> Option<&PlanPhase> {
-    phases
-        .iter()
-        .find(|p| matches!(p.status, PlanStatus::InProgress))
-}
-
-/// Record a plan phase status change to the history log.
-pub fn record_history(
-    project_root: &Path,
-    phase_id: &str,
-    old_status: &PlanStatus,
-    new_status: &PlanStatus,
-) -> anyhow::Result<()> {
-    // Validate state-machine transition. Log a warning for illegal moves;
-    // return an error when strict_transitions is enabled in [plan] config.
-    if !old_status.is_valid_transition_to(new_status) {
-        tracing::warn!(
-            phase = %phase_id,
-            from = %old_status,
-            to = %new_status,
-            "invalid plan phase transition — expected pending→in_progress, \
-             in_progress→done, or in_progress→pending"
-        );
-        // Check strict mode from workflow config.
-        let wf_path = project_root.join(".ta/workflow.toml");
-        let wf_config = TaWorkflowConfig::load_or_default(&wf_path);
-        if wf_config.plan.strict_transitions {
-            anyhow::bail!(
-                "Phase {}: invalid state transition {} → {} (strict_transitions enabled). \
-                 Legal: pending → in_progress → done, or in_progress → pending on reset.",
-                phase_id,
-                old_status,
-                new_status
-            );
-        }
-    }
-
-    let ta_dir = project_root.join(".ta");
-    std::fs::create_dir_all(&ta_dir)?;
-    let history_path = ta_dir.join("plan_history.jsonl");
-
-    let entry = serde_json::json!({
-        "timestamp": chrono::Utc::now().to_rfc3339(),
-        "phase_id": phase_id,
-        "old_status": old_status.to_string(),
-        "new_status": new_status.to_string(),
-    });
-
-    use std::io::Write;
-    let mut file = std::fs::OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(&history_path)?;
-    writeln!(file, "{}", entry)?;
-    Ok(())
-}
-
-/// Mark a phase as `in_progress` in the source PLAN.md.
-///
-/// Called by `ta run --phase <id>` immediately after staging is created,
-/// before the agent launches. Writes to the **source** PLAN.md so that
-/// `ta plan status` reflects active work immediately.
-///
-/// Logs the transition to `.ta/plan_history.jsonl`. No-ops if PLAN.md
-/// doesn't exist or the phase is not found.
-pub fn mark_phase_in_source(project_root: &Path, phase_id: &str) -> anyhow::Result<()> {
-    let plan_path = project_root.join("PLAN.md");
-    if !plan_path.exists() {
-        return Ok(());
-    }
-    let content = std::fs::read_to_string(&plan_path)?;
-    let phases = parse_plan(&content);
-    let old_status = phases
-        .iter()
-        .find(|p| phase_ids_match(&p.id, phase_id))
-        .map(|p| p.status.clone())
-        .unwrap_or(PlanStatus::Pending);
-
-    // Only update if the phase is currently pending (don't downgrade done→in_progress).
-    if !matches!(old_status, PlanStatus::Pending) {
-        return Ok(());
-    }
-
-    let updated = update_phase_status(&content, phase_id, PlanStatus::InProgress);
-    if updated == content {
-        // Phase not found or content unchanged — silently no-op.
-        return Ok(());
-    }
-    std::fs::write(&plan_path, &updated)?;
-    let _ = record_history(project_root, phase_id, &old_status, &PlanStatus::InProgress);
-    Ok(())
-}
-
-/// Reset a phase from `in_progress` back to `pending` in the source PLAN.md.
-///
-/// Called on `ta draft deny`, `ta draft close`, and `ta goal delete` when the
-/// associated goal had a linked plan phase. Logs the transition to
-/// `.ta/plan_history.jsonl` with the provided `note`.
-///
-/// No-ops if the phase is not currently `in_progress`. In particular, a `done`
-/// phase is an explicit, defended invariant here (not just an accident of the
-/// match falling through): a stale/duplicate goal referencing already-completed,
-/// already-merged work must never revert it to pending (v0.17.0.12.11).
-///
-/// Returns `true` if the phase was actually reset, `false` if this was a no-op —
-/// callers use this to avoid printing a misleading "reset to pending" message
-/// when nothing changed.
+/// Reset a phase from `in_progress` back to `pending` in the source PLAN.md,
+/// then best-effort release the daemon's in-memory claim over HTTP so the
+/// next `ta run` doesn't need a daemon restart. Thin wrapper: the core
+/// reset+history-log logic lives in `ta_plan::reset_phase_if_in_progress`;
+/// the daemon-release HTTP call is CLI-process-specific and stays here.
 pub fn reset_phase_if_in_progress(
     project_root: &Path,
     phase_id: &str,
     note: &str,
 ) -> anyhow::Result<bool> {
-    let plan_path = project_root.join("PLAN.md");
-    if !plan_path.exists() {
-        return Ok(false);
+    let did_reset = ta_plan::reset_phase_if_in_progress(project_root, phase_id, note)?;
+    if did_reset {
+        release_daemon_phase_claim(project_root, phase_id);
     }
-    let content = std::fs::read_to_string(&plan_path)?;
-    let phases = parse_plan(&content);
-    let current_status = phases
-        .iter()
-        .find(|p| phase_ids_match(&p.id, phase_id))
-        .map(|p| p.status.clone());
-
-    match current_status {
-        Some(PlanStatus::InProgress) => {}
-        Some(PlanStatus::Done) => return Ok(false), // never revert a completed phase
-        _ => return Ok(false),                      // pending/deferred/missing — nothing to reset
-    }
-
-    let updated = update_phase_status(&content, phase_id, PlanStatus::Pending);
-    if updated == content {
-        return Ok(false);
-    }
-    std::fs::write(&plan_path, &updated)?;
-
-    // Log with a note field appended to the standard history entry.
-    let ta_dir = project_root.join(".ta");
-    std::fs::create_dir_all(&ta_dir)?;
-    let history_path = ta_dir.join("plan_history.jsonl");
-    let entry = serde_json::json!({
-        "timestamp": chrono::Utc::now().to_rfc3339(),
-        "phase_id": phase_id,
-        "old_status": "in_progress",
-        "new_status": "pending",
-        "note": note,
-    });
-    use std::io::Write as _;
-    let mut file = std::fs::OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(&history_path)?;
-    writeln!(file, "{}", entry)?;
-
-    // Best-effort: tell the daemon to release its in-memory claim so future
-    // `ta run` calls succeed without requiring a daemon restart.
-    release_daemon_phase_claim(project_root, phase_id);
-
-    Ok(true)
+    Ok(did_reset)
 }
-
 /// Fire-and-forget HTTP call to release the daemon's in-memory phase claim.
 ///
 /// The daemon holds phase claims in RAM — resetting PLAN.md alone leaves a
@@ -1650,20 +946,8 @@ fn plan_reset(config: &GatewayConfig, phase_id: &str) -> anyhow::Result<()> {
     Ok(())
 }
 
-/// Load plan history entries from the JSONL file.
-pub fn load_history(project_root: &Path) -> anyhow::Result<Vec<serde_json::Value>> {
-    let history_path = project_root.join(".ta/plan_history.jsonl");
-    if !history_path.exists() {
-        return Ok(Vec::new());
-    }
-    let content = std::fs::read_to_string(&history_path)?;
-    let entries: Vec<serde_json::Value> = content
-        .lines()
-        .filter(|line| !line.trim().is_empty())
-        .filter_map(|line| serde_json::from_str(line).ok())
-        .collect();
-    Ok(entries)
-}
+// v0.17.11.1: extracted to ta-plan.
+pub use ta_plan::load_history;
 
 /// Build a suggested `ta run` command for the next pending phase.
 pub fn suggest_next_goal_command(phase: &PlanPhase) -> String {
@@ -1685,7 +969,11 @@ fn detect_schema_from_content(content: &str, source: &str) -> PlanSchema {
         return schema;
     }
 
-    // Fallback: generic ## heading pattern.
+    // Fallback: generic ## heading pattern. `statuses`/`doc_search_dirs` come
+    // from the default schema via struct-update syntax rather than calling
+    // ta-plan's default_statuses()/default_doc_search_dirs() directly, to
+    // avoid importing them outside test code (see the #[cfg(test)] import
+    // above).
     PlanSchema {
         source: source.to_string(),
         phase_patterns: vec![PhasePattern {
@@ -1693,8 +981,7 @@ fn detect_schema_from_content(content: &str, source: &str) -> PlanSchema {
             id_capture: "heading_text".to_string(),
         }],
         status_marker: r"<!--\s*status:\s*(\w+)\s*-->".to_string(),
-        statuses: default_statuses(),
-        doc_search_dirs: default_doc_search_dirs(),
+        ..PlanSchema::default_schema()
     }
 }
 
@@ -2123,63 +1410,8 @@ fn show_status(config: &GatewayConfig, json_output: bool, kpi_output: bool) -> a
     Ok(())
 }
 
-/// Collect dependency warnings for all phases whose declared `depends_on` phases are not Done.
-pub fn collect_dependency_warnings(phases: &[PlanPhase]) -> Vec<String> {
-    let mut warnings = Vec::new();
-    for phase in phases {
-        if phase.depends_on.is_empty() {
-            continue;
-        }
-        for dep_id in &phase.depends_on {
-            let dep_done = phases
-                .iter()
-                .any(|p| phase_ids_match(&p.id, dep_id) && p.status == PlanStatus::Done);
-            if !dep_done {
-                warnings.push(format!(
-                    "Phase {} depends on {} which is not yet done.",
-                    phase.id, dep_id,
-                ));
-            }
-        }
-    }
-    warnings
-}
-
-/// Returns the binary version string at compile time.
-pub fn binary_version() -> &'static str {
-    env!("CARGO_PKG_VERSION")
-}
-
-/// Returns true if the phase ID is a sub-phase (has 4 or more numeric components).
-///
-/// A sub-phase has the form `vX.Y.Z.N` (or deeper like `vX.Y.Z.N.M`), as opposed
-/// to a top-level phase `vX.Y.Z`. Non-semver IDs like "4b" or "Phase 0" are never
-/// considered sub-phases.
-pub fn is_sub_phase(id: &str) -> bool {
-    let stripped = match id.strip_prefix('v') {
-        Some(s) if s.starts_with(|c: char| c.is_ascii_digit()) => s,
-        _ => return false,
-    };
-    stripped.split('.').count() >= 4
-}
-
-/// Returns the immediate parent phase ID for a sub-phase, or `None` if the ID
-/// is not a sub-phase.
-///
-/// Drops only the last dot-separated component — `v0.16.0.1` → `Some("v0.16.0")`,
-/// `v0.15.30.5.1` → `Some("v0.15.30.5")`. Previously this always truncated to
-/// exactly 3 components regardless of depth (`v0.15.30.5.1` → `"v0.15.30"`),
-/// which for deeply nested IDs (`v0.17.0.12.30` → `"v0.17.0"`) skipped over the
-/// real immediate parent (`v0.17.0.12`) and landed on an unrelated, often
-/// long-done, top-level ancestor instead (v0.17.0.12.30 fix).
-pub fn parent_phase_id(id: &str) -> Option<String> {
-    if !is_sub_phase(id) {
-        return None;
-    }
-    let stripped = id.strip_prefix('v').unwrap_or(id);
-    let parts: Vec<&str> = stripped.split('.').collect();
-    Some(format!("v{}", parts[..parts.len() - 1].join(".")))
-}
+// v0.17.11.1: extracted to ta-plan.
+pub use ta_plan::{binary_version, collect_dependency_warnings, is_sub_phase, parent_phase_id};
 
 /// Find the phase matching `parent_id`, but only if it is still relevant
 /// context — i.e. not already `Done`.
@@ -2195,301 +1427,11 @@ fn find_active_parent_phase<'a>(phases: &'a [PlanPhase], parent_id: &str) -> Opt
         .find(|p| phase_ids_match(&p.id, parent_id) && p.status != PlanStatus::Done)
 }
 
-/// Parse a semver-style phase ID like "v0.14.3" or "v0.13.17.1" into a comparable tuple of u32s.
-///
-/// Only phases whose ID starts with `v` followed by digits are considered.
-/// Returns `None` for non-semver IDs (e.g., "4b", "Phase 1").
-fn parse_semver_id(id: &str) -> Option<Vec<u32>> {
-    let stripped = id.strip_prefix('v')?;
-    // Must start with a digit after the 'v'
-    if !stripped.starts_with(|c: char| c.is_ascii_digit()) {
-        return None;
-    }
-    let parts: Option<Vec<u32>> = stripped.split('.').map(|s| s.parse::<u32>().ok()).collect();
-    parts
-}
-
-/// Maximum number of sub-phase segments (beyond `major.minor.patch`) that are
-/// encoded losslessly as dot-separated semver pre-release identifiers.
-///
-/// Chosen because this project's deepest observed phase nesting (`v0.17.0.12.10`,
-/// i.e. 2 sub-segments) is comfortably under it, while still bounding the
-/// pre-release chain length. Beyond this depth, `phase_id_to_semver` truncates
-/// and appends a build-metadata discriminator instead (see below).
-const MAX_SUBPHASE_DEPTH: usize = 4;
-
-/// Derive a stable 4-hex-char discriminator from a phase ID string.
-///
-/// Deterministic per phase-id (not per-commit) so the same overflowing phase ID
-/// always produces the same build-metadata suffix, regardless of which commit
-/// happens to perform the bump.
-fn phase_id_discriminator(phase_id: &str) -> String {
-    use sha2::Digest as _;
-    let digest = sha2::Sha256::digest(phase_id.as_bytes());
-    format!("{:x}", digest)[..4].to_string()
-}
-
-/// Convert a plan phase ID to the canonical workspace semver string.
-///
-/// Phase ID mapping (per CLAUDE.md version policy):
-///   v0.14.22           → "0.14.22-alpha"
-///   v0.14.22.1         → "0.14.22-alpha.1"
-///   v0.14.22.2         → "0.14.22-alpha.2"
-///   v0.15.0            → "0.15.0-alpha"
-///   v0.17.0.12.9       → "0.17.0-alpha.12.9"
-///   v0.17.0.12.9.1     → "0.17.0-alpha.12.9.1"
-///
-/// Sub-phase IDs are handled generically: any number of segments beyond
-/// `major.minor.patch` are appended in order as dot-separated pre-release
-/// identifiers, up to `MAX_SUBPHASE_DEPTH` of them. This is valid semver —
-/// the spec allows pre-release identifiers to repeat indefinitely — and
-/// Cargo's semver parser accepts it.
-///
-/// Beyond `MAX_SUBPHASE_DEPTH` sub-segments, the pre-release chain is truncated
-/// to the first `MAX_SUBPHASE_DEPTH` and a build-metadata suffix (`+xxxx`) is
-/// appended containing a 4-hex-char discriminator derived from a stable hash of
-/// the *full* phase ID. Semver build metadata is explicitly excluded from
-/// precedence/ordering comparisons by spec-compliant tooling (cargo, the
-/// `semver` crate, shields.io), so this never breaks version sorting or the
-/// `version-check` CI job's equality check — it's purely an extra disambiguator
-/// for the rare case of very deep nesting.
-///
-/// Non-semver phase IDs (e.g., "4b", "Phase 1") return `None` — no auto-bump.
-pub fn phase_id_to_semver(phase_id: &str) -> Option<String> {
-    let parts = parse_semver_id(phase_id)?;
-    if parts.len() < 3 {
-        return None;
-    }
-    let (major, minor, patch) = (parts[0], parts[1], parts[2]);
-    let sub_segments = &parts[3..];
-
-    if sub_segments.is_empty() {
-        return Some(format!("{}.{}.{}-alpha", major, minor, patch));
-    }
-
-    if sub_segments.len() <= MAX_SUBPHASE_DEPTH {
-        let pre_release = sub_segments
-            .iter()
-            .map(|s| s.to_string())
-            .collect::<Vec<_>>()
-            .join(".");
-        return Some(format!(
-            "{}.{}.{}-alpha.{}",
-            major, minor, patch, pre_release
-        ));
-    }
-
-    // Beyond MAX_SUBPHASE_DEPTH: truncate the pre-release chain and disambiguate
-    // with a build-metadata discriminator derived from the full phase ID.
-    let truncated = sub_segments[..MAX_SUBPHASE_DEPTH]
-        .iter()
-        .map(|s| s.to_string())
-        .collect::<Vec<_>>()
-        .join(".");
-    let discriminator = phase_id_discriminator(phase_id);
-    Some(format!(
-        "{}.{}.{}-alpha.{}+{}",
-        major, minor, patch, truncated, discriminator
-    ))
-}
-
-/// Loudly warn (tracing + stderr) that a phase ID could not be converted to a
-/// version via `phase_id_to_semver` and that the version bump was skipped.
-///
-/// This is the fix for the 6-phase version drift root cause: `phase_id_to_semver`
-/// returning `None` used to be handled with a silent `if let Some(...) = ...`
-/// that skipped the bump with no visibility at all. Call this from the `else`
-/// arm everywhere `phase_id_to_semver` gates an auto-bump.
-pub fn warn_unparseable_phase_id_for_bump(phase_id: &str) {
-    tracing::warn!(
-        phase = %phase_id,
-        "phase_id_to_semver could not parse phase ID — version bump skipped"
-    );
-    eprintln!(
-        "[version] Warning: phase ID '{}' could not be converted to a version \
-         (expected `vMAJOR.MINOR.PATCH[.SUB...]`, e.g. `v0.17.0` or `v0.17.0.12.9`) — \
-         version bump skipped. If this phase should bump the version, rename it to \
-         follow that format, or run ./scripts/bump-version.sh manually.",
-        phase_id
-    );
-}
-
-/// Check for out-of-order phases: a `Done` phase appears after a `Pending` phase
-/// in document order (for phases with semver-style IDs only).
-///
-/// Returns deduplicated human-readable warning strings: one line per pending phase
-/// showing the count of later-done phases (v0.15.19.4.2 deduplication).
-pub fn check_phase_order(phases: &[PlanPhase]) -> Vec<String> {
-    // Collect (index, id, status) for semver phases only.
-    let semver_phases: Vec<(usize, &PlanPhase)> = phases
-        .iter()
-        .enumerate()
-        .filter(|(_, p)| parse_semver_id(&p.id).is_some())
-        .collect();
-
-    // pending_ids_in_order: insertion-ordered list of pending phase IDs
-    // pending_later_done: parallel counts of Done phases appearing after each pending phase
-    let mut pending_ids_in_order: Vec<String> = Vec::new();
-    let mut pending_later_done: Vec<usize> = Vec::new();
-
-    for (_, phase) in &semver_phases {
-        if phase.status == PlanStatus::Pending {
-            pending_ids_in_order.push(phase.id.clone());
-            pending_later_done.push(0);
-        } else if phase.status == PlanStatus::Done {
-            // Count this Done phase against all currently-seen Pending phases.
-            for count in pending_later_done.iter_mut() {
-                *count += 1;
-            }
-        }
-    }
-
-    // Emit one line per pending phase that has later-done violations.
-    pending_ids_in_order
-        .iter()
-        .zip(pending_later_done.iter())
-        .filter_map(|(pid, &count)| {
-            if count == 0 {
-                return None;
-            }
-            Some(format!(
-                "[warn] {} is still pending — {} later phase(s) are complete (out of order)",
-                pid, count
-            ))
-        })
-        .collect()
-}
-
-/// Detect phases that have no `<!-- status: ... -->` marker in PLAN.md content.
-///
-/// Returns a list of phase IDs that are missing a status marker.
-/// These phases parse as `Pending` due to the `find_status_in_lookahead` fallback,
-/// which may produce false "pending" counts in `ta plan status`.
-pub fn detect_missing_status_markers(content: &str) -> Vec<String> {
-    use regex::Regex;
-
-    let status_re = match Regex::new(r"<!--\s*status:\s*\w+\s*-->") {
-        Ok(r) => r,
-        Err(_) => return vec![],
-    };
-
-    // Phase header patterns (same as default schema).
-    let header_patterns: &[&str] = &[
-        r"^###\s+(v[\d]+\.[\d]+\.[\d]+(?:\.[\d]+)?)\s+[—\-]",
-        r"^##\s+Phase\s+([\w.]+)\s+[—\-]",
-        r"^###\s+(v[\d]+\.[\d]+)\s+[—\-]",
-    ];
-    let compiled: Vec<_> = header_patterns
-        .iter()
-        .filter_map(|p| Regex::new(p).ok())
-        .collect();
-
-    let lines: Vec<&str> = content.lines().collect();
-    let mut missing = Vec::new();
-
-    for (i, line) in lines.iter().enumerate() {
-        let trimmed = line.trim();
-        let mut matched_id: Option<String> = None;
-        for pat in &compiled {
-            if let Some(caps) = pat.captures(trimmed) {
-                matched_id = caps.get(1).map(|m| m.as_str().to_string());
-                break;
-            }
-        }
-        if let Some(id) = matched_id {
-            // Check if next non-empty line has a status marker.
-            let next = lines.get(i + 1).map(|l| l.trim()).unwrap_or("");
-            if !status_re.is_match(next) {
-                missing.push(id);
-            }
-        }
-    }
-
-    missing
-}
-
-/// Scan PLAN.md for phases where all items are `[x]` but status marker is not `done`.
-///
-/// Returns `(phase_id, line_number_of_header)` pairs.
-pub fn find_phases_needing_done_marker(content: &str) -> Vec<(String, usize)> {
-    use regex::Regex;
-
-    let schema = PlanSchema::default_schema();
-    let phases = parse_plan_with_schema(content, &schema);
-    let missing_markers = detect_missing_status_markers(content);
-    let missing_set: std::collections::HashSet<&str> =
-        missing_markers.iter().map(|s| s.as_str()).collect();
-
-    let mut result = Vec::new();
-    let lines: Vec<&str> = content.lines().collect();
-
-    // Phase header detection (same patterns).
-    let header_patterns: &[&str] = &[
-        r"^###\s+(v[\d]+\.[\d]+\.[\d]+(?:\.[\d]+)?)\s+[—\-]",
-        r"^##\s+Phase\s+([\w.]+)\s+[—\-]",
-        r"^###\s+(v[\d]+\.[\d]+)\s+[—\-]",
-    ];
-    let compiled_re = Regex::new(r"<!--\s*status:\s*(\w+)\s*-->").ok();
-
-    for phase in &phases {
-        // Only flag if all plan items are checked.
-        if phase.status == PlanStatus::Done {
-            continue; // Already marked done.
-        }
-        if !missing_set.contains(phase.id.as_str()) && phase.status != PlanStatus::Pending {
-            continue; // Has a non-done status marker — user intent.
-        }
-        // Find the header line for this phase.
-        let header_line_idx = lines.iter().position(|l| {
-            let trimmed = l.trim();
-            header_patterns.iter().any(|p| {
-                Regex::new(p)
-                    .ok()
-                    .and_then(|r| r.captures(trimmed))
-                    .map(|caps| caps.get(1).map(|m| m.as_str()) == Some(phase.id.as_str()))
-                    .unwrap_or(false)
-            })
-        });
-        let _ = compiled_re.as_ref(); // suppress unused warning
-        if let Some(idx) = header_line_idx {
-            result.push((phase.id.clone(), idx + 1));
-        }
-    }
-
-    result
-}
-
-/// Check whether the binary version is ahead of the last completed phase.
-///
-/// Returns `Some(warning)` if the binary is ahead, `None` if in sync.
-///
-/// Resolves "last completed phase" via [`last_completed_phase_id`] (the
-/// dependency-graph-based computation) rather than its own document-position
-/// scan — consolidated in v0.17.0.12.30 to avoid a second, independent
-/// heuristic for the same question living alongside the canonical one.
-pub fn check_version_sync(phases: &[PlanPhase]) -> Option<String> {
-    let last_done_id = last_completed_phase_id(phases);
-    let highest_phase = phases
-        .iter()
-        .find(|p| phase_ids_match(&p.id, &last_done_id) && p.status == PlanStatus::Done)?;
-    let binary = binary_version();
-
-    // Compare binary version vs highest sequential done phase.
-    // Parse both as semver tuples. Strip pre-release suffixes from binary version.
-    let binary_base = binary.split('-').next().unwrap_or(binary);
-    let binary_parts = parse_semver_id(&format!("v{}", binary_base))?;
-    let phase_parts = parse_semver_id(&highest_phase.id)?;
-
-    if binary_parts > phase_parts {
-        Some(format!(
-            "Binary version ({}) is ahead of highest sequential completed phase ({}). \
-             Consider pinning for release — see CLAUDE.md 'Public Release Process'.",
-            binary, highest_phase.id,
-        ))
-    } else {
-        None
-    }
-}
+// v0.17.11.1: extracted to ta-plan.
+pub use ta_plan::{
+    check_phase_order, check_version_sync, detect_missing_status_markers,
+    find_phases_needing_done_marker, phase_id_to_semver, warn_unparseable_phase_id_for_bump,
+};
 
 fn show_next(config: &GatewayConfig, filter: Option<&str>) -> anyhow::Result<()> {
     let phases = load_plan(&config.workspace_root)?;
@@ -4639,206 +3581,11 @@ pub fn create_gap_semver(last_done: &str, existing_phases: &[PlanPhase]) -> Stri
     }
 }
 
-/// Compute the set of phases that are actually ready to be worked on next:
-/// status `Pending` (not `Done`, not already claimed `InProgress`, not
-/// `Deferred`), and every phase declared in that phase's own `**Depends on**`
-/// line is `Done`.
-///
-/// This is the single source of truth for "what's next" — never a
-/// document-position heuristic. A phase's position in PLAN.md (especially
-/// after a merge) does not reflect real completion or dependency order; only
-/// its own declared dependencies do (v0.17.0.12.30).
-pub fn next_actionable_phases(phases: &[PlanPhase]) -> Vec<&PlanPhase> {
-    phases
-        .iter()
-        .filter(|p| p.status == PlanStatus::Pending)
-        .filter(|p| {
-            p.depends_on.iter().all(|dep_id| {
-                phases
-                    .iter()
-                    .find(|d| phase_ids_match(&d.id, dep_id))
-                    .is_some_and(|d| d.status == PlanStatus::Done)
-            })
-        })
-        .collect()
-}
-
-/// Partition a candidate set of pending phases into dependency waves
-/// (v0.17.0.12.34) — every phase in wave N depends only on phases in waves
-/// < N (already done, by construction of the candidate set) or on other
-/// candidates that come earlier in the wave order; phases within the same
-/// wave declare no ordering or API-impact conflict between them and are
-/// safe to run concurrently.
-///
-/// Read-only analysis: this does not launch anything. Callers typically
-/// pass [`next_actionable_phases`]'s result (or a specific human-selected
-/// batch — this phase's parallel execution is opt-in, not a global
-/// always-on default) to see what's safe to parallelize before choosing to
-/// actually run a batch concurrently.
-///
-/// Dependencies pointing at phases outside the candidate set (e.g. an
-/// already-`Done` phase) are treated as already satisfied and dropped
-/// before graph construction — only intra-batch ordering matters here.
-pub fn candidate_waves(phases: &[PlanPhase]) -> Result<Vec<Vec<String>>, String> {
-    let ids: std::collections::HashSet<&str> = phases.iter().map(|p| p.id.as_str()).collect();
-    let nodes: Vec<ta_workflow::WaveNode> = phases
-        .iter()
-        .map(|p| {
-            let deps: Vec<String> = p
-                .depends_on
-                .iter()
-                .filter(|d| ids.contains(d.as_str()) || ids.iter().any(|id| phase_ids_match(id, d)))
-                .cloned()
-                .collect();
-            ta_workflow::WaveNode::new(p.id.clone())
-                .with_deps(deps)
-                .with_impact_tags(p.api_impact.clone())
-        })
-        .collect();
-
-    ta_workflow::plan_waves(&nodes).map_err(|e| e.to_string())
-}
-
-// Ascending comparator: real semver IDs compare by value; a phase with a
-// non-semver ID always sorts after every semver phase (never picked by
-// `max_by` over a real dependency, never mistaken for the primary
-// next-actionable target over a real one).
-fn by_semver(a: &&PlanPhase, b: &&PlanPhase) -> std::cmp::Ordering {
-    match (parse_semver_id(&a.id), parse_semver_id(&b.id)) {
-        (Some(sa), Some(sb)) => sa.cmp(&sb),
-        (Some(_), None) => std::cmp::Ordering::Greater,
-        (None, Some(_)) => std::cmp::Ordering::Less,
-        (None, None) => std::cmp::Ordering::Equal,
-    }
-}
-
-/// The single primary next-actionable phase ID, if any.
-///
-/// Consolidates the "pick one" selection out of [`next_actionable_phases`]'s
-/// result set (v0.17.0.12.30 item 4: one shared function, not a second
-/// ad-hoc scan) — takes the lowest-semver phase among those whose
-/// dependencies are all satisfied, preferring phases with a real semver ID
-/// and falling back to the first actionable phase if none parse. Returns
-/// `None` when there is no next-actionable phase — e.g. the whole plan is
-/// done (build-milestone automation treats this as "no-op, milestone
-/// complete", not an error).
-pub fn next_actionable_phase_id(phases: &[PlanPhase]) -> Option<String> {
-    let actionable = next_actionable_phases(phases);
-    actionable
-        .iter()
-        .copied()
-        .filter(|p| parse_semver_id(&p.id).is_some())
-        .min_by(by_semver)
-        .or_else(|| actionable.first().copied())
-        .map(|p| p.id.clone())
-}
-
-/// Find the last completed phase ID, used for gap-semver generation and the
-/// `ta plan status` version-check line.
-///
-/// Derives this from the *dependency graph*, not document position: takes the
-/// lowest-semver phase from [`next_actionable_phases`] (the immediate next
-/// step, via [`next_actionable_phase_id`]) and returns the highest-semver
-/// phase among *that phase's own* declared dependencies — i.e. what
-/// specifically had to finish for the next step to become ready. All version
-/// comparisons go through [`parse_semver_id`], never a document-position scan.
-///
-/// Falls back to the highest-semver `Done` phase overall (still compared via
-/// `parse_semver_id`, never by position) when there is no next-actionable
-/// phase or it has no declared dependencies — e.g. the whole plan is done, or
-/// a project with no `**Depends on**` lines at all.
-pub fn last_completed_phase_id(phases: &[PlanPhase]) -> String {
-    let primary_id = next_actionable_phase_id(phases);
-    let primary = primary_id
-        .as_deref()
-        .and_then(|id| phases.iter().find(|p| p.id == id));
-
-    if let Some(phase) = primary {
-        let last_dep = phase
-            .depends_on
-            .iter()
-            .filter_map(|dep_id| phases.iter().find(|d| phase_ids_match(&d.id, dep_id)))
-            .filter(|d| d.status == PlanStatus::Done)
-            .max_by(by_semver);
-        if let Some(dep) = last_dep {
-            return dep.id.clone();
-        }
-    }
-
-    phases
-        .iter()
-        .filter(|p| p.status == PlanStatus::Done && parse_semver_id(&p.id).is_some())
-        .max_by(by_semver)
-        .map(|p| p.id.clone())
-        .unwrap_or_else(|| "v0.0.0".to_string())
-}
-
-/// Insert an ad-hoc phase stub into PLAN.md immediately after the last Done phase.
-///
-/// The stub has `<!-- status: in_progress -->` since it starts immediately.
-/// If the phase ID already exists, this is a no-op.
-pub fn insert_adhoc_phase(project_root: &Path, phase_id: &str, title: &str) -> anyhow::Result<()> {
-    let plan_path = project_root.join("PLAN.md");
-    if !plan_path.exists() {
-        return Ok(());
-    }
-    let content = std::fs::read_to_string(&plan_path)?;
-
-    // No-op if phase already exists.
-    if content.contains(phase_id) {
-        return Ok(());
-    }
-
-    // Find the last Done phase and insert after its block.
-    // Simple heuristic: find the last `<!-- status: done -->` line, then insert
-    // after the paragraph following it (next blank line after the line).
-    let stub = format!(
-        "\n### {} — {}\n<!-- status: in_progress -->\n*Inserted goal — not in original plan.*\n",
-        phase_id, title
-    );
-
-    // Find insertion point: after the last `<!-- status: done -->` section.
-    // We walk backward looking for the last occurrence of "status: done", then
-    // find the next blank line after it (end of that phase's intro paragraph).
-    let insert_pos = find_insert_pos_after_last_done(&content);
-    let updated = format!(
-        "{}{}{}",
-        &content[..insert_pos],
-        stub,
-        &content[insert_pos..]
-    );
-    std::fs::write(&plan_path, &updated)?;
-    Ok(())
-}
-
-/// Find the character position immediately after the last Done phase block.
-fn find_insert_pos_after_last_done(content: &str) -> usize {
-    // Find the last "<!-- status: done -->" occurrence.
-    let done_marker = "<!-- status: done -->";
-    let last_done_pos = content.rfind(done_marker);
-    let Some(done_start) = last_done_pos else {
-        // No done phases — insert at the end.
-        return content.len();
-    };
-
-    // From done_start, scan forward to find the end of this phase's content block.
-    // End of block = the next `### ` or `## ` header, or end of file.
-    let after_done = done_start + done_marker.len();
-    let rest = &content[after_done..];
-
-    // Look for the next section header.
-    for (i, line) in rest.lines().enumerate() {
-        let trimmed = line.trim();
-        if i > 0 && (trimmed.starts_with("### ") || trimmed.starts_with("## ")) {
-            // Insert before this header.
-            let byte_offset: usize = rest.lines().take(i).map(|l| l.len() + 1).sum();
-            return after_done + byte_offset;
-        }
-    }
-
-    // No next header found — insert at end.
-    content.len()
-}
+// v0.17.11.1: extracted to ta-plan.
+pub use ta_plan::{
+    candidate_waves, insert_adhoc_phase, last_completed_phase_id, next_actionable_phase_id,
+    next_actionable_phases,
+};
 
 /// Auto-detect the plan phase for a goal run.
 ///
