+++ crates/ta-plan/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "ta-plan"
+version.workspace = true
+edition = "2021"
+description = "PLAN.md schema, parsing, and phase-tracking logic for Trusted Autonomy (extracted from ta-cli, v0.17.11.1)"
+license = "Apache-2.0"
+repository = "https://github.com/trustedautonomy/ta"
+homepage = "https://github.com/trustedautonomy/ta"
+keywords = ["ai", "agent", "autonomy", "plan"]
+categories = ["development-tools"]
+
+[dependencies]
+anyhow = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+serde_yaml = { workspace = true }
+regex = { workspace = true }
+chrono = { workspace = true }
+sha2 = { workspace = true }
+tracing = { workspace = true }
+uuid = { workspace = true }
+ta-goal = { path = "../ta-goal", version = "0.17.10-alpha.4" }
+ta-submit = { path = "../ta-submit", version = "0.17.10-alpha.4" }
+ta-workflow = { path = "../ta-workflow", version = "0.17.10-alpha.4" }
+
+[dev-dependencies]
+tempfile = { workspace = true }
